workspace structure rework
